pe-static-analyzer/requirements.txt
# Core analysis
lief>=0.13.0
pefile>=2023.2.7
yara-python>=4.3.1

# Database
sqlalchemy>=2.0.0

# CLI
rich>=13.0.0
click>=8.1.0

# GUI
PySide6>=6.5.0

# Reporting
jinja2>=3.1.0
reportlab>=4.0.0

# Utilities
python-magic>=0.4.27
colorama>=0.4.6
pyyaml>=6.0
requests>=2.31.0
capstone>=5.0.1
graphviz>=0.20.1
